requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiolimiter==1.2.1
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
autopep8==2.3.2
beautifulsoup4==4.13.4
binance==0.3.43
ccxt==4.4.85
ccxt_download==0.4.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.3
curl_cffi==0.11.1
cvxpy==1.6.5
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
ecos==2.0.14
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
osqp==1.0.4
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==5.24.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.8.0
pycodestyle==2.13.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pyportfolioopt==1.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
scs==3.2.7.post2
seaborn==0.13.2
setuptools==80.8.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vectorbt==0.27.3
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
yahoo-finance==1.4.0
yarl==1.20.0
yfinance==0.2.61
